+++ R/summarizeMilestoneTime.R
@@ -39,8 +39,8 @@
 #'
 #' trial$add_arms(sample_ratio = c(1, 1), control, active)
 #'
-#' action_at_final <- function(trial, milestone_name){
-#'   locked_data <- trial$get_locked_data(milestone_name)
+#' action_at_final <- function(trial){
+#'   locked_data <- trial$get_locked_data('final analysis')
 #'   fitLogrank(Surv(PFS, PFS_event) ~ arm, placebo = 'control arm',
 #'              data = locked_data, alternative = 'less')
 #'   invisible(NULL)