+++ tests/testthat/test-trial.R
@@ -198,16 +198,16 @@ test_that('filters are supported when defining milestones', {
 
   }
 
-  interim_action <- function(trial, milestone_name){
+  interim_action <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('interim')
     trial$save(value = locked_data %>% filter(dll3 == 'high' & x > -2) %>% nrow(),
                name = 'n_at_lock')
   }
 
-  final_action <- function(trial, milestone_name){
+  final_action <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
     tmp1 <- locked_data %>% filter(dll3 == 'high' & x > -2)
     trial$save(value = sum(tmp1$pfs_event), name = 'n_pfs_high')
     trial$save(value = sum(locked_data$pfs_event), name = 'n_pfs_all')
@@ -278,7 +278,7 @@ test_that('endpoint event counts work as expected when duration is adapted', {
 
   trial$add_arms(sample_ratio = c(1, 2), pbo, trt)
 
-  action_at_interim <- function(trial, milestone_name){
+  action_at_interim <- function(trial){
     trial$set_duration(duration = 40)
   }
 
@@ -384,7 +384,7 @@ test_that('endpoint event counts work as expected when an arm is removed', {
 
   interim1 <- milestone(name = 'interim1',
                         when = eventNumber(endpoint = 'or', n = 200),
-                        action = function(trial, milestone_name){trial$remove_arms('trt1')})
+                        action = function(trial){trial$remove_arms('trt1')})
 
   interim2 <- milestone(name = 'interim2',
                         when = eventNumber(endpoint = 'pfs', n = 240) &
@@ -491,7 +491,7 @@ test_that('custom data can be re-used in multiple trials', {
   trial$save_custom_data(value = list(x = 1, y = 'a'), name = 'config')
 
   final <- milestone(name = 'final',
-                     action = function(trial, milestone_name) {
+                     action = function(trial) {
                        trial$get_custom_data('config')},
                      when = calendarTime(time = 40))
 
@@ -520,9 +520,9 @@ test_that('trial data can be replicated', {
 
   trial$add_arms(sample_ratio = 1, pbo)
 
-  act <- function(trial, milestone_name){
+  act <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
     trial$save(value = median(locked_data$ep), name = 'median')
     trial$save(value = mean(locked_data$ep), name = 'mean')
     trial$save(value = sd(locked_data$ep), name = 'sd')
@@ -595,9 +595,9 @@ test_that('fitLinear works as expected', {
 
   trial$add_arms(sample_ratio = c(1, 2), pbo, trt)
 
-  act <- function(trial, milestone_name){
+  act <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
 
     n <- nrow(locked_data)
     locked_data$covar1 <- rnorm(n)
@@ -655,9 +655,9 @@ test_that('fitLinear can compute ATE as expected in additive model', {
 
   trial$add_arms(sample_ratio = c(1, 2), pbo, trt)
 
-  act <- function(trial, milestone_name){
+  act <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
 
     n <- nrow(locked_data)
     locked_data$covar1 <- rnorm(n)
@@ -718,9 +718,9 @@ test_that('fitLogistic can compute ATE as expected in model without covariates',
 
   trial$add_arms(sample_ratio = c(1, 2), pbo, trt)
 
-  act <- function(trial, milestone_name){
+  act <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
 
     trial$save(value = nrow(locked_data), name = 'n')
 
@@ -839,9 +839,9 @@ test_that('fitLogistic can compute regression coefficient as expected in model w
 
   trial$add_arms(sample_ratio = c(1, 2), pbo, trt)
 
-  act <- function(trial, milestone_name){
+  act <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
     locked_data$x <- rnorm(nrow(locked_data))
     locked_data$y <- rnorm(nrow(locked_data))
     locked_data$z <- rbinom(nrow(locked_data), 1, .5)
@@ -912,9 +912,9 @@ test_that('fitLogrank works as expected', {
 
   trial$add_arms(sample_ratio = c(1, 2), pbo, trt)
 
-  act <- function(trial, milestone_name){
+  act <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
 
     fit <- fitLogrank(Surv(ep, ep_event) ~ arm, placebo = 'pbo', data = locked_data, alternative = 'less')
 
@@ -967,9 +967,9 @@ test_that('fitCoxph can compute main effect of arm', {
 
   trial$add_arms(sample_ratio = c(1, 2), pbo, trt)
 
-  act <- function(trial, milestone_name){
+  act <- function(trial){
 
-    locked_data <- trial$get_locked_data(milestone_name)
+    locked_data <- trial$get_locked_data('final')
 
     n <- nrow(locked_data)
     locked_data$covar1 <- rnorm(n)
@@ -1045,9 +1045,9 @@ test_that('sample ratio can be updated to switch between permuted block and samp
   trial$add_arms(sample_ratio = c(1, 1, 1, 1), pbo, low, med, high)
 
 
-  action1 <- function(trial, milestone_name){
+  action1 <- function(trial){
 
-    dat1 <- trial$get_locked_data(milestone_name)
+    dat1 <- trial$get_locked_data('interim1')
 
     prob <- prop.table(trial$get_sample_ratio())
 
@@ -1063,10 +1063,10 @@ test_that('sample ratio can be updated to switch between permuted block and samp
                         action = action1
   )
 
-  action2 <- function(trial, milestone_name){
+  action2 <- function(trial){
 
     dat1 <- trial$get_locked_data('interim1')
-    dat2 <- trial$get_locked_data(milestone_name) %>%
+    dat2 <- trial$get_locked_data('interim2') %>%
       filter(!(patient_id %in% dat1$patient_id))
 
     prob <- prop.table(trial$get_sample_ratio())
@@ -1083,10 +1083,10 @@ test_that('sample ratio can be updated to switch between permuted block and samp
                         action = action2
   )
 
-  action3 <- function(trial, milestone_name){
+  action3 <- function(trial){
 
     dat2 <- trial$get_locked_data('interim2')
-    dat3 <- trial$get_locked_data(milestone_name) %>%
+    dat3 <- trial$get_locked_data('interim3') %>%
       filter(!(patient_id %in% dat2$patient_id))
 
     prob <- prop.table(trial$get_sample_ratio())
@@ -1103,10 +1103,10 @@ test_that('sample ratio can be updated to switch between permuted block and samp
   )
 
 
-  action4 <- function(trial, milestone_name){
+  action4 <- function(trial){
 
     dat3 <- trial$get_locked_data('interim3')
-    dat4 <- trial$get_locked_data(milestone_name) %>%
+    dat4 <- trial$get_locked_data('final') %>%
       filter(!(patient_id %in% dat3$patient_id))
 
     prob <- prop.table(trial$get_sample_ratio())